add server side validation

+++ src/main/java/com/medinar/fullstackspringbootreact/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.medinar.fullstackspringbootreact.student;
 
+import com.medinar.fullstackspringbootreact.student.exception.BadRequestException;
+import com.medinar.fullstackspringbootreact.student.exception.NotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,13 +16,22 @@ public class StudentService {
         return studentRepository.findAll();
     }
 
-    public void addStudent(Student student) {
-        // check if email is taken
+    public void addStudent(Student student) throws BadRequestException {
+        Boolean emailExists = studentRepository.existsByEmail(student.getEmail());
+        if (emailExists) {
+            throw new BadRequestException(String.format(
+                    "Email `%s` already exist",
+                    student.getEmail()
+            ));
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
-        // check if student exist
+        if(!studentRepository.existsById(studentId)) {
+            throw new NotFoundException(
+                    "Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
